+++ NEWS.md
@@ -6,6 +6,9 @@
 - feat: add auto-refresh on TODO edits.
 - refactor!(CLI): make add and remove task first class commands.
   Use `tsk add/remove` instead of `tsk task add/remove`.
+- feat: add annotate command to add a note.
+- feat: add hash to notes so that they can be removed with the `tsk remove` command.
+- feat: Make the pop up window's content (task) editable.
 
 ## 0.1.0 (2026-05-28)